Create Leap Motion Fingers
In this part of the code, the leap motion will be able to detect the arm and individual fingers. By taking this, it will understand how the arms and fingers are aligned and where they are located. Maybe I can get it to give me exact coordinates of the arms and fingers.